Annex functions

Index

MessyTimeSeries.centred_moving_average

MessyTimeSeries.check_bounds

MessyTimeSeries.companion_form

MessyTimeSeries.demean

MessyTimeSeries.error_info

MessyTimeSeries.forward_backwards_rw_interpolation

MessyTimeSeries.interpolate_series

MessyTimeSeries.is_vector_in_matrix

MessyTimeSeries.isconverged

MessyTimeSeries.lag

MessyTimeSeries.mean_skipmissing

MessyTimeSeries.nan_to_missing!

MessyTimeSeries.soft_thresholding

MessyTimeSeries.solve_discrete_lyapunov

MessyTimeSeries.standardise

MessyTimeSeries.std_skipmissing

MessyTimeSeries.sum_skipmissing

MessyTimeSeries.trimmed_mean

MessyTimeSeries.verb_message

Functions

MessyTimeSeries.centred_moving_average — Method

centred_moving_average(X::Union{FloatMatrix, JMatrix{Float64}}, n::Int64, T::Int64, window::Int64)

Compute the centred moving average of X.

Arguments

X: observed measurements (nxT)

n and T are the number of series and observations

window is the total number of observations (lagging, current and leading) included in the average

MessyTimeSeries.check_bounds — Method

check_bounds(X::Real, LB::Real, UB::Real)

Check whether X is larger or equal than LB and lower or equal than UB

check_bounds(X::Real, LB::Real)

Check whether X is larger or equal than LB

MessyTimeSeries.companion_form — Method

companion_form(Ψ::AbstractArray{Float64,2}; extended::Bool=false)

Construct the companion form matrix from the generic coefficients Ψ.

If extended is true, it increases the typical dimension of the companion matrix by n rows.

MessyTimeSeries.demean — Method

demean(X::FloatVector)
demean(X::FloatMatrix)

Demean complete data.

demean(X::JVector{Float64})
demean(X::JMatrix{Float64})

Demean incomplete data.

Examples

julia> demean([1.0; 1.5; 2.0; 2.5; 3.0])
5-element Array{Float64,1}:
 -1.0
 -0.5
  0.0
  0.5
  1.0

julia> demean([1.0 3.5 1.5 4.0 2.0; 4.5 2.5 5.0 3.0 5.5])
2×5 Array{Float64,2}:
 -1.4   1.1  -0.9   1.6  -0.4
  0.4  -1.6   0.9  -1.1   1.4

MessyTimeSeries.error_info — Method

error_info(err::Exception)
error_info(err::RemoteException)

Return error main information

MessyTimeSeries.forward_backwards_rw_interpolation — Method

forward_backwards_rw_interpolation(X::JMatrix{Float64}, n::Int64, T::Int64)

Interpolate each non-stationary series in X, in turn, using a random walk logic both forward and backwards in time.

Arguments

X: observed measurements (nxT)

n and T are the number of series and observations

forwardbackwardsrw_interpolation(X::FloatMatrix, n::Int64, T::Int64)

Return X.

MessyTimeSeries.interpolate_series — Method

interpolate_series(X::JMatrix{Float64}, n::Int64, T::Int64)

Interpolate each series in X, in turn, by replacing missing observations with the sample average of the non-missing values.

Arguments

X: observed measurements (nxT)

n and T are the number of series and observations

interpolate_series(X::FloatMatrix, n::Int64, T::Int64)

Return X.

Arguments

X: observed measurements (nxT)

n and T are the number of series and observations

MessyTimeSeries.is_vector_in_matrix — Method

is_vector_in_matrix(vect::IntVector, matr::IntMatrix)
is_vector_in_matrix(vect::FloatVector, matr::FloatMatrix)

Check whether the vector vect is included in the matrix matr.

Examples

julia> isvectorin_matrix([1; 2], [1 2; 2 3]) true

MessyTimeSeries.isconverged — Method

isconverged(new::Float64, old::Float64, tol::Float64, ε::Float64, increasing::Bool)

Check whether new is close enough to old.

Arguments

new: new objective or loss

old: old objective or loss

tol: tolerance

ε: small Float64

increasing: true if new increases, at each iteration, with respect to old

MessyTimeSeries.lag — Method

lag(X::FloatArray, p::Int64)

Construct the data required to run a standard vector autoregression.

Arguments

X: observed measurements (nxT), where n and T are the number of series and observations.

p: number of lags in the vector autoregression

Output

X_{t}

X_{t-1}

MessyTimeSeries.mean_skipmissing — Method

mean_skipmissing(X::AbstractVector{Float64})
mean_skipmissing(X::AbstractVector{Union{Missing, Float64}})

Compute the mean of the observed values in X.

mean_skipmissing(X::AbstractMatrix{Float64})
mean_skipmissing(X::AbstractMatrix{Union{Missing, Float64}})

Compute the mean of the observed values in X column wise.

Examples

julia> mean_skipmissing([1.0; missing; 3.0])
2.0

julia> mean_skipmissing([1.0 2.0; missing 3.0; 3.0 5.0])
3-element Array{Float64,1}:
 1.5
 3.0
 4.0

MessyTimeSeries.nan_to_missing! — Method

nan_to_missing!(X::JArray{Float64})

Replace NaN with missing in X.

MessyTimeSeries.soft_thresholding — Method

soft_thresholding(z::Float64, ζ::Float64)

Soft thresholding operator.

MessyTimeSeries.solve_discrete_lyapunov — Method

solve_discrete_lyapunov(A::AbstractArray{Float64,2}, Q::SymMatrix)

Use a bilinear transformation to convert the discrete Lyapunov equation to a continuous Lyapunov equation, which is then solved using BLAS.

The notation used for representing the discrete Lyapunov equation is

$P - APA' = Q$,

where P and Q are symmetric. This equation is transformed into

B'P + PB = -C

References

Kailath (1980, page 180)

MessyTimeSeries.standardise — Method

standardise(X::FloatVector)
standardise(X::FloatMatrix)

Standardise complete data.

standardise(X::JVector{Float64})
standardise(X::JMatrix{Float64})

Standardise incomplete data.

Examples

julia> standardise([1.0; 1.5; 2.0; 2.5; 3.0])
5-element Array{Float64,1}:
 -1.2649110640673518
 -0.6324555320336759
  0.0
  0.6324555320336759
  1.2649110640673518

julia> standardise([1.0 3.5 1.5 4.0 2.0; 4.5 2.5 5.0 3.0 5.5])
2×5 Array{Float64,2}:
 -1.08173    0.849934  -0.695401   1.23627   -0.309067
  0.309067  -1.23627    0.695401  -0.849934   1.08173

MessyTimeSeries.std_skipmissing — Method

std_skipmissing(X::AbstractVector{Float64})
std_skipmissing(X::AbstractVector{Union{Missing, Float64}})

Compute the standard deviation of the observed values in X.

std_skipmissing(X::AbstractMatrix{Float64})
std_skipmissing(X::AbstractMatrix{Union{Missing, Float64}})

Compute the standard deviation of the observed values in X column wise.

Examples

julia> std_skipmissing([1.0; missing; 3.0])
1.4142135623730951

julia> std_skipmissing([1.0 2.0; missing 3.0; 3.0 5.0])
3-element Array{Float64,1}:
   0.7071067811865476
 NaN
   1.4142135623730951

MessyTimeSeries.sum_skipmissing — Method

sum_skipmissing(X::AbstractVector{Float64})
sum_skipmissing(X::AbstractVector{Union{Missing, Float64}})

Compute the sum of the observed values in X.

sum_skipmissing(X::AbstractMatrix{Float64})
sum_skipmissing(X::AbstractMatrix{Union{Missing, Float64}})

Compute the sum of the observed values in X column wise.

Examples

julia> sum_skipmissing([1.0; missing; 3.0])
4.0

julia> sum_skipmissing([1.0 2.0; missing 3.0; 3.0 5.0])
3-element Array{Float64,1}:
 3.0
 3.0
 8.0

MessyTimeSeries.trimmed_mean — Method

trimmed_mean(X::AbstractArray{Float64,1}, α::Float64)

Compute the trimmed mean of X (i.e., the sample average of X having removed its α smallest and largest values).

MessyTimeSeries.verb_message — Method

verb_message(verb::Bool, message::String)

@info message if verb is true.